requirements.txt
alembic
fastapi
uvicorn
psycopg2
python-decouple
sqlmodel